# BACKEND_CODE/Swagger_Api_Doc/swagger.yaml
openapi: 3.0.3
info:
  title: VideoTube - The Production Ready Application
  description: >-
    # VideoTube API Documentation

    Welcome to the API documentation for VideoTube, a production-ready application developed by CoderSwarup. This API serves as a key component in your journey to mastering backend development by guiding you through the creation of a YouTube-like video-sharing platform. Explore the rich features and gain hands-on experience in building a robust and scalable backend system.

    ## Features

    - **Production Ready:** VideoTube is meticulously designed to be production-ready, ensuring reliability, scalability, and efficiency in handling user registrations. Through this API, you'll learn best practices in building a resilient backend infrastructure.

    ## 🎓 Learning Backend Development with VideoTube

    Embark on a comprehensive learning experience in backend development as you delve into the creation of a YouTube application clone. The API documentation provides insights into the following key areas:

    - **GitHub Repository:** Access the complete source code and project setup instructions on [GitHub](https://github.com/CoderSwarup). Follow the provided installation instructions to set up VideoTube on your local environment.

...

...

    ## 🏁 Installation

...

    ## 🐞 Reporting Issues

...

    ## Contact

...

  version: 1.0.0
  contact: {}
servers:
  - url: 'http://localhost:3000/api/v1'
paths:
...
  /users/get-user-details:
    get:
      summary: Get details of the currently logged-in user
      description: Retrieve details of the user currently authenticated through JWT.
      tags:
        - Users
      security:
        - cookieAuth: [] # Assuming you have a security definition named 'cookieAuth'
      responses:
        '200':
          description: User details fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized, user not authenticated
          content:
            application/json:
              example:
                error: Unauthorized
      operationId: getCurrentUser

  # Update User Details
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatar:
          type: string
        coverImage:
          type: string
        watchHistory:
          type: array
          items:
            type: string # Assuming Video IDs are string
        password:
          type: string
        refreshToken:
          type: string
      required:
        - username
        - email
        - fullName
        - avatar
        - password

    # Response Schema
...